fix: sudden EOF while connecting through NPP
The problem is - both client and server successfully connects to each other,
so there are at least 2 different connections.
The server must serve both of them with further dropping the hanging one.
The client must return the first one and close another.
So...
If we are a server - we just put successful connection in a queue and return
on next "Accept" called.
If we are a client - do as we did before.